add base optimizer for future InferenceOptimizer keras
